+++ src/engine/script.c
@@ -13,12 +13,10 @@ struct script
 	duk_uarridx_t id;
 };
 
-static script_t* script_from_js_function (void* heapptr);
-
-static int       s_next_script_id = 0;
+static int s_next_script_id = 0;
 
 void
-initialize_scripts(void)
+scripts_init(void)
 {
 	console_log(1, "initializing JS script manager");
 	
@@ -29,13 +27,13 @@ initialize_scripts(void)
 }
 
 void
-shutdown_scripts(void)
+scripts_uninit(void)
 {
 	console_log(1, "shutting down JS script manager");
 }
 
 bool
-evaluate_script(const char* filename, bool as_module)
+script_eval(const char* filename, bool as_module)
 {
 	sfs_file_t*    file = NULL;
 	path_t*        path = NULL;
@@ -84,7 +82,7 @@ on_error:
 }
 
 script_t*
-compile_script(const lstring_t* source, const char* fmt_name, ...)
+script_new(const lstring_t* source, const char* fmt_name, ...)
 {
 	va_list    ap;
 	lstring_t* name;
@@ -99,7 +97,7 @@ compile_script(const lstring_t* source, const char* fmt_name, ...)
 	
 	// this wouldn't be necessary if Duktape duk_get_heapptr() gave us a strong reference.
 	// instead we get this ugliness where the compiled function is saved in the global stash
-	// so it doesn't get eaten by the garbage collector. yummy!
+	// so it doesn't get eaten by the garbage collector.  yummy!
 	duk_push_global_stash(g_duk);
 	duk_get_prop_string(g_duk, -1, "scripts");
 	duk_push_lstring_t(g_duk, source);
@@ -112,18 +110,39 @@ compile_script(const lstring_t* source, const char* fmt_name, ...)
 	lstr_free(name);
 
 	script->id = s_next_script_id++;
-	return ref_script(script);
+	return script_ref(script);
 }
 
 script_t*
-ref_script(script_t* script)
+script_new_func(duk_context* ctx, duk_idx_t idx)
+{
+	script_t* script;
+
+	idx = duk_require_normalize_index(ctx, idx);
+	duk_require_function(ctx, idx);
+	
+	if (!(script = calloc(1, sizeof(script_t))))
+		return NULL;
+
+	duk_push_global_stash(g_duk);
+	duk_get_prop_string(g_duk, -1, "scripts");
+	duk_dup(g_duk, idx);
+	duk_put_prop_index(g_duk, -2, s_next_script_id);
+	duk_pop_2(g_duk);
+
+	script->id = s_next_script_id++;
+	return script_ref(script);
+}
+
+script_t*
+script_ref(script_t* script)
 {
 	++script->refcount;
 	return script;
 }
 
 void
-free_script(script_t* script)
+script_free(script_t* script)
 {
 	if (script == NULL || --script->refcount > 0)
 		return;
@@ -140,7 +159,7 @@ free_script(script_t* script)
 }
 
 void
-run_script(script_t* script, bool allow_reentry)
+script_run(script_t* script, bool allow_reentry)
 {
 	bool was_in_use;
 
@@ -148,7 +167,7 @@ run_script(script_t* script, bool allow_reentry)
 		return;
 	
 	// check whether an instance of the script is already running.
-	// if it is, but the script is reentrant, allow it. otherwise, return early
+	// if it is, but the script is reentrant, allow it.  otherwise, return early
 	// to prevent multiple invocation.
 	if (script->is_in_use && !allow_reentry) {
 		console_log(3, "skipping execution of script #%u, already in use", script->id);
@@ -158,11 +177,11 @@ run_script(script_t* script, bool allow_reentry)
 
 	console_log(4, "executing script #%u", script->id);
 
-	// ref the script in case it gets freed during execution. the owner
+	// ref the script in case it gets freed during execution.  the owner
 	// may be destroyed in the process and we don't want to end up crashing.
-	ref_script(script);
+	script_ref(script);
 	
-	// get the compiled script from the stash and run it. so dumb...
+	// get the compiled script from the stash and run it.  so dumb...
 	script->is_in_use = true;
 	duk_push_global_stash(g_duk);
 	duk_get_prop_string(g_duk, -1, "scripts");
@@ -171,48 +190,5 @@ run_script(script_t* script, bool allow_reentry)
 	duk_pop_3(g_duk);
 	script->is_in_use = was_in_use;
 
-	free_script(script);
-}
-
-script_t*
-duk_require_sphere_script(duk_context* ctx, duk_idx_t index, const char* name)
-{
-	lstring_t* codestring;
-	script_t*  script;
-
-	index = duk_require_normalize_index(ctx, index);
-
-	if (duk_is_callable(ctx, index)) {
-		// caller passed function directly
-		script = script_from_js_function(duk_get_heapptr(ctx, index));
-	}
-	else if (duk_is_string(ctx, index)) {
-		// caller passed code string, compile it
-		codestring = duk_require_lstring_t(ctx, index);
-		script = compile_script(codestring, "%s", name);
-		lstr_free(codestring);
-	}
-	else if (duk_is_null_or_undefined(ctx, index))
-		return NULL;
-	else
-		duk_error_ni(ctx, -1, DUK_ERR_TYPE_ERROR, "script must be string, function, or null/undefined");
-	return script;
-}
-
-static script_t*
-script_from_js_function(void* heapptr)
-{
-	script_t* script;
-
-	if (!(script = calloc(1, sizeof(script_t))))
-		return NULL;
-
-	duk_push_global_stash(g_duk);
-	duk_get_prop_string(g_duk, -1, "scripts");
-	duk_push_heapptr(g_duk, heapptr);
-	duk_put_prop_index(g_duk, -2, s_next_script_id);
-	duk_pop_2(g_duk);
-
-	script->id = s_next_script_id++;
-	return ref_script(script);
+	script_free(script);
 }