+++ src/engine/script.h
@@ -3,14 +3,13 @@
 
 typedef struct script script_t;
 
-void             initialize_scripts (void);
-void             shutdown_scripts   (void);
-bool             evaluate_script    (const char* filename, bool as_module);
-script_t*        compile_script     (const lstring_t* script, const char* fmt_name, ...);
-script_t*        ref_script         (script_t* script);
-void             free_script        (script_t* script);
-void             run_script         (script_t* script, bool allow_reentry);
-
-script_t* duk_require_sphere_script (duk_context* ctx, duk_idx_t index, const char* name);
+void      scripts_init    (void);
+void      scripts_uninit  (void);
+bool      script_eval     (const char* filename, bool as_module);
+script_t* script_new      (const lstring_t* script, const char* fmt_name, ...);
+script_t* script_new_func (duk_context* ctx, duk_idx_t idx);
+script_t* script_ref      (script_t* script);
+void      script_free     (script_t* script);
+void      script_run      (script_t* script, bool allow_reentry);
 
 #endif // MINISPHERE__SCRIPT_H__INCLUDED